+++ acesso_movel/snakRestaurant/app/src/main/java/amsi/dei/estg/ipleiria/snakrestaurant/models/Shopping_card.java
@@ -5,14 +5,14 @@ public class Shopping_card {
     private Integer id_shopping;
     private  long id_product_shopping;
     private String name_shopping;
-    private int price_shopping;
+    private double price_shopping;
     public int id_category_shopping;
     public int quantidade_shopping;
 
     //ID_PRODUCT_SHOPPING, NAME_SHOPPING, PRICE_SHOPPING, ID_CATEGORY_SHOPPING, QUANTIDADE
 
 
-    public Shopping_card( long id_product_shopping, String name_shopping, int price_shopping, int id_category_shopping, int quantidade_shopping) {
+    public Shopping_card( long id_product_shopping, String name_shopping, double price_shopping, int id_category_shopping, int quantidade_shopping) {
 
         this.id_product_shopping = id_product_shopping;
         this.name_shopping = name_shopping;
@@ -21,7 +21,7 @@ public class Shopping_card {
         this.quantidade_shopping = quantidade_shopping;
     }
 
-    public Shopping_card(int id_shopping, long id_product_shopping, String name_shopping, int price_shopping, int id_category_shopping, int quantidade_shopping) {
+    public Shopping_card(int id_shopping, long id_product_shopping, String name_shopping, double price_shopping, int id_category_shopping, int quantidade_shopping) {
         this.id_shopping = id_shopping;
         this.id_product_shopping = id_product_shopping;
         this.name_shopping = name_shopping;
@@ -54,11 +54,11 @@ public class Shopping_card {
         this.name_shopping = name_shopping;
     }
 
-    public int getPrice_shopping() {
+    public double getPrice_shopping() {
         return price_shopping;
     }
 
-    public void setPrice_shopping(int price_shopping) {
+    public void setPrice_shopping(double price_shopping) {
         this.price_shopping = price_shopping;
     }
